WIP getting the SDK tests working
# Conflicts:
#	tests/util/provisioner.go

# Conflicts:
#	tests/util/provisioner.go


Refactored directory creation in Makefile

The Makefile has been updated to use the 'addprefix' function for creating directories. This change simplifies the code and improves readability.

Updated YAML configuration

The properties of the 'tee' resource in the YAML file have been modified. The 'stdin' property has been replaced with an 'args' property, which now contains a 'content' sub-property. This change improves the flexibility and readability of the configuration.

Regen examples

Enhanced security with client certificates

Added the generation of client certificates to enhance security in the .NET SDK test suite. This involves creating a new certificate bundle before initializing the provisioner. Also, removed an unused import from the provisioner utility file.


Refactoring and restructuring to maybe make it work

# Conflicts:
#	Makefile

Cleanup and instrumentation

Regen examples

Add make dep to tests