Refactor certificate generation in tests
The process of generating certificates for testing has been refactored. The responsibility of creating the certificates has been moved from the lifecycle test to the provider suite setup. This change simplifies the lifecycle test and makes it easier to manage certificates across different tests. Additionally, a new method 'WithCerts' was introduced in TestHost interface replacing 'CreateCertBundle'.