Added builder struct and methods in cmd package
A new file, builder.go, has been added to the cmd package. This file introduces a 'builder' struct with an array of strings as its field. Three methods have been implemented for this struct: 'op', 'opv', and 'add'. The 'op' method adds a name to the args if input is true. The 'opv' method adds a name-value pair to the args if value is not empty. The 'add' method appends parts to the args but panics if more than two parts are passed.