Add wget resource (#89)
* Add wget bin to proto

* Added builder struct and methods in cmd package

A new file, builder.go, has been added to the cmd package. This file introduces a 'builder' struct with an array of strings as its field. Three methods have been implemented for this struct: 'op', 'opv', and 'add'. The 'op' method adds a name to the args if input is true. The 'opv' method adds a name-value pair to the args if value is not empty. The 'add' method appends parts to the args but panics if more than two parts are passed.

* Refactored TeeArgs struct and related methods

The TeeArgs struct and its associated methods have been refactored for better organization. The Cmd() and ExpectedFiles() methods are now defined immediately after the struct, improving readability. Error messages in Create(), Update(), and Delete() functions have also been updated to provide more specific context.

* Updated KeyPem field to be secret

The KeyPem field in the Config struct has been updated to be treated as a secret by the provider. This change enhances security by ensuring sensitive data is not exposed.

* Added support for rm and wget commands

The provisioner service now supports the 'rm' and 'wget' commands. This update extends the switch case in the bin function to handle these additional cases, improving the versatility of command handling.

* Add wget resource

* Add wget test

* Fixin up stuff

Except now something is actually wrong

* Configuring ginkgo a little better

* Go crap

* Wget unit test

* Holy shit finally figured out what it was

* Holy shit finally figured out what it was

* Sus out the problem and log it in the test

* Omg that was dumb

* Remove debugging files

* Fixing stuff

* Regen SDKs