Added 'mv' command functionality
A new file, mv.go, has been added to the cmd package. This file includes the implementation of a 'mv' command with various options such as backup, destination, directory, force and more. The MvArgs struct is defined to hold these options. Two methods Cmd() and ExpectedFiles() are implemented on this struct.

The Mv type is also introduced with Create(), Update(), and Delete() methods implementing respective interfaces from infer package.

Finally, the newly created 'mv' command is registered in provider.go by adding it to the list of inferred resources.