Added test for file move operation
A new test case has been added to the service tests. This test checks the functionality of moving a file from one location to another. It creates a temporary directory and a file within it, moves the file to a new location within the same directory, and then verifies that the move was successful by checking various conditions such as exit code, stdout content, created files list, moved files map and existence of old/new files. The temporary directory is cleaned up after each test run.