Add custom update and delete (#108)
* Initial iml

* Updated command.proto with minor changes

The commit includes a few modifications in the 'command.proto' file. The 'UpdateRequest' message has been updated, replacing 'create' with 'previous'. Similarly, in the 'DeleteRequest', 'create' has been replaced by 'previous'. In addition to these changes, an extra enumeration value, BIN_TOUCH = 9, has been added to the enum Bin.

* Refactored lifecycle tests

Significant changes were made to the lifecycle_test.go file. The ExpectedOutput maps were removed and replaced with Expect functions within the Hook function for each test case. This change simplifies the code by removing redundancy and makes it easier to understand what each test is checking for. Additionally, a new customFile was added, along with corresponding checks in the Hook function. A new operation was also added to update triggers in one of the test cases.

* Refactor provisioner update and delete methods

The code changes refactor the Update and Delete methods in the provisioner package. The logic for skipping during preview has been added to the Update method. Variables 'expectCreated' and 'expectMoved' have been introduced to store expected created and moved files respectively. The Create field in pb.UpdateRequest and pb.DeleteRequest has been replaced with Previous, which is more semantically accurate. Corresponding changes have also been made in service.go and service_test.go files of the provisioner command package.

* Added optional Command to DeleteRequest

An optional field 'Command' has been added to the DeleteRequest message. This allows for more flexibility when handling delete requests, as it provides an option to include a command in the request if necessary.

* Added custom delete functionality

The service now supports a custom delete operation. If a command is provided in the DeleteRequest, it will execute this as a create operation and return the result in the DeleteResponse. This allows for more flexible deletion processes.

* Enhanced command handling and diffing logic

Significant changes include:
- Updated the 'Diff' function to handle 'customUpdate' property differences.
- Added a new method 'UpdateKind' to determine the type of update based on 'CustomUpdate'.
- Introduced a new method 'DeleteBeforeReplace' that checks if there's any custom update before replacing.
- Refined the parsing of commands by switching from binary value checking to explicit case handling for each command.

* Enhanced custom command handling and diffing logic

Significant changes include:
- Added support for custom delete commands in the CommandState
- Updated parsing of custom update and delete commands to use inputs instead of state
- Enhanced property diffing logic in mv.go to use a default kind instead of always using UpdateReplace
- Introduced checks for changes in lifecycle tests, allowing tests to continue even if no changes are detected
- Expanded test coverage with additional scenarios for moving files and executing custom commands.

* Refactor command parsing in update and delete methods

The code changes involve refactoring the way commands are parsed in both the Update and Delete methods. Previously, the parseCommand function was called with 's.CustomUpdate' as an argument. This has been corrected to use 'inputs.CustomUpdate' for the Update method and 's.CustomDelete' for the Delete method, ensuring that the correct custom commands are being parsed.

* Add a bunch of logging and revert the test back to its working state

* Refactor lifecycle tests for better readability

The lifecycle tests have been refactored to improve readability and maintainability. The changes include:
- Replacing the 'resource' package alias with 'pr'
- Adding more detailed test cases for the 'Mv' command
- Simplifying the structure of some test cases by removing unnecessary nesting
- Using context in function calls instead of creating a new background context each time

* Refactored lifecycle tests for command resources

- Simplified the structure of lifecycle tests for Chmod, Mkdir, Mktemp, Rm, Tar, Tee and Wget commands.
- Removed unnecessary BeforeAll blocks and moved their contents into the test cases.
- Consolidated resource declaration and test execution into a single run statement.
- Improved readability by reducing nesting levels.

* Fix linting error

* Regen SDKs